+++ eox_core/edxapp_wrapper/backends/bearer_authentication_j_v1_test.py
@@ -13,3 +13,18 @@ def get_bearer_authentication():
     except ImportError:
         BearerAuthentication = object
     return BearerAuthentication
+
+
+def get_bearer_authentication_allow_inactive_user():  # pylint: disable=invalid-name
+    """Allow to get the class BearerAuthenticationAllowInactiveUser from
+    https://github.com/eduNEXT/edunext-platform/tree/master/openedx/core/lib/api/authentication.py
+
+    Returns:
+        BearerAuthentication function.
+    """
+    try:
+        from openedx.core.lib.api.authentication import BearerAuthenticationAllowInactiveUser
+    except ImportError:
+        BearerAuthenticationAllowInactiveUser = object
+
+    return BearerAuthenticationAllowInactiveUser