+++ eox_core/edxapp_wrapper/bearer_authentication.py
@@ -14,4 +14,13 @@ def get_bearer_authentication():
     return backend.get_bearer_authentication()
 
 
+def get_bearer_authentication_allow_inactive_user():  # pylint: disable=invalid-name
+    """ Gets BearerAuthentication class. """
+    backend_function = settings.EOX_CORE_BEARER_AUTHENTICATION
+    backend = import_module(backend_function)
+
+    return backend.get_bearer_authentication_allow_inactive_user()
+
+
 BearerAuthentication = get_bearer_authentication()  # pylint: disable=invalid-name
+BearerAuthenticationAllowInactiveUser = get_bearer_authentication_allow_inactive_user()  # pylint: disable=invalid-name